test draw credits

+++ SDK/credgem/models.py
@@ -1,13 +1,12 @@
-from decimal import Decimal
 from typing import Dict, List, Optional
 from pydantic import BaseModel
 
 
 class Balance(BaseModel):
     credit_type_id: str
-    available: Decimal
-    held: Decimal
-    spent: Decimal
+    available: float
+    held: float
+    spent: float
 
 
 class WalletResponse(BaseModel):
@@ -32,7 +31,7 @@ class TransactionResponse(BaseModel):
     id: str
     wallet_id: str
     credit_type_id: str
-    amount: Decimal
+    amount: float
     type: str
     description: Optional[str] = None
     issuer: str
